requirements.txt
# Основные зависимости
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.12.0
python-dotenv

# Для логирования и отладки (опционально, но рекомендуется)
colorlog>=6.7.0

# Для тестирования
pytest>=7.0.0
responses>=0.18.0  # для мокинга HTTP-запросов в тестах

# Для расписания запуска (если будешь запускать по расписанию)
schedule>=1.2.0

# Добавьте эти строки
Pillow>=9.0.0
